Split config file to separate files for each class
